% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_pages.R
\name{count_pages}
\alias{count_pages}
\title{Given search query, return how many pages of search results there are}
\usage{
count_pages(htm_input)
}
\arguments{
\item{htm_input}{xml node containing the original html content of the page}
}
\description{
Given search query, return how many pages of search results there are
}
\details{
- The "page" query parameter isn't used on the first page of hits -- this means that page 2 of hits uses the query parameter ?page=1. So need to increment this number by 1.
- Need to handle situation when there are no hits for a search.
- Need to handle situation when there is only one page of hits for a search.
}
