% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_attrs_generic.R
\name{extract_children_generic}
\alias{extract_children_generic}
\title{For a given html node, extract the names and text values of its children.}
\usage{
extract_children_generic(x)
}
\arguments{
\item{x}{xml_node with name/text to extract}
}
\value{
Data frame with one column per html element.
}
\description{
aaa
}
