% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_result_list}
\alias{extract_result_list}
\title{Get first page of search hits for a search term.}
\usage{
extract_result_list(htm_input)
}
\arguments{
\item{query}{character vector of search terms}
}
\value{
Data frame containing the URLs and description excerpts for the first page of search hits.
}
\description{
aaa
}
