% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_search_contents.R
\name{get_search_contents}
\alias{get_search_contents}
\title{Retrieve the first page of search results for a given query.}
\usage{
get_search_contents(solr_search_term, production = TRUE)
}
\arguments{
\item{production}{(logical) perform search on the production site? default is TRUE}

\item{term}{character string containing search term}
}
\value{
List containing the following elements:
- html xml_node containing the search result page's HTML
- search_url character string displaying the query URL
}
\description{
Retrieves the HTML contents from the first page of search hits for a given query.
}
\details{
When using the test server (www-test.library.kent.edu), needed to
modify approach because of the SSL certificate.
- http://stackoverflow.com/questions/34551299/how-to-pass-ssl-verifypeer-in-rvest
- https://curl.haxx.se/docs/sslcerts.html
}
