% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_search_url.R
\name{get_search_url}
\alias{get_search_url}
\title{Return the URL corresponding to a query using Kent State University Library's site search.}
\usage{
get_search_url(term = "", production = TRUE)
}
\arguments{
\item{term}{character string containing search term}

\item{logical}{perform search on the production site? default is TRUE}
}
\value{
Character vector of search URLs.
}
\description{
Returns URL corresponding to a chosen query string.
}
\examples{
get_search_url("hours")
get_search_url("hours", production=FALSE)
}
