% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_dataframe.R
\name{result_dataframe}
\alias{result_dataframe}
\title{Get first page of search hits for a search term.}
\usage{
result_dataframe(nodelist)
}
\arguments{
\item{nodelist}{xml node where direct children are all li.search-result elements}
}
\value{
Data frame containing the title text, actual link, search description, and displayed URL.
}
\description{
aaa
}
\details{
- Need to handle situation when there are no hits for a search.
}
