% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ksu_search.R
\name{run_ksu_search}
\alias{run_ksu_search}
\title{Test of one or more search queries on either the live site or test site.}
\usage{
run_ksu_search(search_terms, production = TRUE)
}
\arguments{
\item{production}{(logical) perform search on the production site? default is TRUE}

\item{term}{character vector containing search terms}
}
\value{
Dataframe containing the first page of hits for each search query.
}
\description{
Run a test of one or more search queries, using either the live (production) site or the test server.
}
